FDG-6361 Seperated tests
